Move shape testing code to pkg/util/test
Want to reuse for testing other types as well.
