+++ pkg/mimirpb/query_response_extra_test.go
@@ -3,7 +3,6 @@
 package mimirpb
 
 import (
-	"fmt"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -12,7 +11,8 @@ import (
 	"github.com/grafana/regexp"
 	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/stretchr/testify/require"
-	"github.com/xlab/treeprint"
+
+	"github.com/grafana/mimir/pkg/util/test"
 )
 
 func TestAllPrometheusStatusValues(t *testing.T) {
@@ -86,73 +86,5 @@ func TestFloatHistogramProtobufTypeRemainsInSyncWithPrometheus(t *testing.T) {
 	protoType := reflect.TypeOf(FloatHistogram{})
 	prometheusType := reflect.TypeOf(histogram.FloatHistogram{})
 
-	requireSameShape(t, prometheusType, protoType)
-}
-
-func requireSameShape(t *testing.T, expectedType reflect.Type, actualType reflect.Type) {
-	expectedFormatted := prettyPrintType(expectedType)
-	actualFormatted := prettyPrintType(actualType)
-
-	require.Equal(t, expectedFormatted, actualFormatted)
-}
-
-func prettyPrintType(t reflect.Type) string {
-	if t.Kind() != reflect.Struct {
-		panic(fmt.Sprintf("expected %s to be a struct but is %s", t.Name(), t.Kind()))
-	}
-
-	tree := treeprint.NewWithRoot("<root>")
-	addTypeToTree(t, tree)
-
-	return tree.String()
-}
-
-func addTypeToTree(t reflect.Type, tree treeprint.Tree) {
-	if t.Kind() != reflect.Struct {
-		panic(fmt.Sprintf("unexpected kind %s", t.Kind()))
-	}
-
-	for i := 0; i < t.NumField(); i++ {
-		f := t.Field(i)
-
-		switch f.Type.Kind() {
-		case reflect.Slice:
-			name := fmt.Sprintf("+%v %s: []%s", f.Offset, f.Name, f.Type.Elem().Kind())
-
-			if isPrimitive(f.Type.Elem().Kind()) {
-				tree.AddNode(name)
-			} else {
-				addTypeToTree(f.Type.Elem(), tree.AddBranch(name))
-			}
-		default:
-			name := fmt.Sprintf("+%v %s: %s", f.Offset, f.Name, f.Type.Kind())
-			tree.AddNode(name)
-		}
-	}
-}
-
-func isPrimitive(k reflect.Kind) bool {
-	switch k {
-	case reflect.Bool,
-		reflect.Int,
-		reflect.Int8,
-		reflect.Int16,
-		reflect.Int32,
-		reflect.Int64,
-		reflect.Uint,
-		reflect.Uint8,
-		reflect.Uint16,
-		reflect.Uint32,
-		reflect.Uint64,
-		reflect.Uintptr,
-		reflect.Float32,
-		reflect.Float64,
-		reflect.Complex64,
-		reflect.Complex128,
-		reflect.String,
-		reflect.UnsafePointer:
-		return true
-	default:
-		return false
-	}
+	test.RequireSameShape(t, prometheusType, protoType)
 }
